separate tracked objects class into new file
