+++ skelly_tracker/trackers/bright_point_tracker/brightest_point_tracker.py
@@ -4,7 +4,8 @@ from typing import Dict
 import cv2
 import numpy as np
 
-from skelly_tracker.trackers.base_tracker.base_tracker import BaseTracker, TrackedObject
+from skelly_tracker.trackers.base_tracker.base_tracker import BaseTracker
+from skelly_tracker.trackers.base_tracker.tracked_object import TrackedObject
 
 UPPER_BOUND_COLOR = [255, 255, 255]
